+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## next
 
+- Added `ViewModel#enforceScheduledRenders()` to immediately execute scheduled renders
+- Changed `ViewModel#scheduleRender()` to use `setTimeout()` instead of `Promise.resolve()` to ensure proper processing of event loop tasks, eliminating unnecessary renders
 - Fixed `ViewModel#setPageParams()` to normalize the `hash` by ensuring it starts with `#`, preventing unnecessary events; for example, passing `#page` and `page` into the method will now consistently result in `#page` being stored in `ViewModel#pageHash`
 
 ## 1.0.0-beta.90 (15-11-2024)